refactoring etl job and script
