+++ dcpquery/etl/__init__.py
@@ -2,11 +2,12 @@ import os, tempfile, logging
 
 from dcplib.etl import DSSExtractor
 
+from dcpquery.db.materialized_views import create_materialized_view_tables
 from dcpquery.etl.load import BundleLoader
 from dcpquery.etl.transform import transform_bundle
 
 from .. import config
-from ..db import Bundle, File, BundleFileLink, DCPMetadataSchemaType
+from ..db import Bundle, File, BundleFileLink
 
 logger = logging.getLogger(__name__)
 
@@ -46,57 +47,6 @@ def update_process_join_table():
 def dcpquery_etl_finalizer(extractor):
     create_materialized_view_tables()
     update_process_join_table()
-
-
-def update_bundles_materialized_view():
-    config.db_session.execute(
-        """
-        REFRESH MATERIALIZED VIEW CONCURRENTLY bundles
-        """
-    )
-
-
-def update_files_materialized_view():
-    config.db_session.execute(
-        """
-        REFRESH MATERIALIZED VIEW CONCURRENTLY files
-        """
-    )
-
-
-def create_dcp_schema_type_materialized_views(matviews):
-    schema_types = [schema[0] for schema in
-                    config.db_session.query(DCPMetadataSchemaType).with_entities(DCPMetadataSchemaType.name).all()]
-    for schema_type in schema_types:
-        if schema_type not in matviews:
-            config.db_session.execute(
-                f"""
-                  CREATE MATERIALIZED VIEW {schema_type} AS
-                  SELECT f.* FROM files as f
-                  WHERE f.dcp_schema_type_name = '{schema_type}'
-                """
-            )
-            config.db_session.execute(
-                f"""
-                CREATE UNIQUE INDEX IF NOT EXISTS {schema_type+'_idx'} ON {schema_type} (fqid);
-
-                """
-            )
-        else:
-            config.db_session.execute(
-                f"""
-                REFRESH MATERIALIZED VIEW CONCURRENTLY {schema_type}
-                """
-            )
-
-
-def create_materialized_view_tables():
-    matviews = [x[0] for x in config.db_session.execute("SELECT matviewname FROM pg_catalog.pg_matviews;").fetchall()]
-    config.reset_db_timeout_seconds(880)
-    update_bundles_materialized_view()
-    update_files_materialized_view()
-    create_dcp_schema_type_materialized_views(matviews)
-    config.db_session.commit()
 
 
 def etl_one_bundle(bundle_uuid, bundle_version):
