+++ scripts/etl.sh
@@ -11,7 +11,7 @@ cd /mnt
 export AWS_DEFAULT_REGION=us-east-1 DEBIAN_FRONTEND=noninteractive LC_ALL=C.UTF-8 LANG=C.UTF-8
 source /etc/profile
 
-apt-get install -qqy --no-install-suggests --no-install-recommends jq moreutils gettext make virtualenv zip unzip httpie git xz-utils dstat
+apt-get install -qqy --no-install-suggests --no-install-recommends jq moreutils gettext make virtualenv zip unzip httpie git xz-utils
 
 virtualenv --python=python3 .venv
 source .venv/bin/activate
@@ -19,10 +19,10 @@ rm -rf query-service
 export GIT_DISCOVERY_ACROSS_FILESYSTEM=1
 git clone https://github.com/HumanCellAtlas/query-service.git
 cd query-service
-git checkout dunitz-projects-migration-2
+#git checkout BRANCH_NAME # uncomment out to use a branch other than master
 pip install --quiet -r requirements-dev.txt
 
-source environment
+source environment # set to env scripts/launch_etl_job.sh called in
 #export DSS_HOST='manually set url"
 S3_CACHE_URL=s3://${SERVICE_S3_BUCKET}/etl/cache.xz
 aws s3 cp $S3_CACHE_URL - | tar -xJ || true
